💄 Changed the view switcher's style in CommonView
